separate prefs and long-session UI tests
